[any message]

+++ app/server.go
@@ -55,8 +55,8 @@ func parseRequest(conn net.Conn, buff []byte) {
 	conn.Read(inputRequestData)
 	binary.BigEndian.PutUint32(buff[:4], 0)                                                    // Message Size
 	binary.BigEndian.PutUint32(buff[4:8], binary.BigEndian.Uint32(inputRequestData[8:12]))     //correlationid
-	binary.BigEndian.PutUint16(buff[8:10], uint16(validateApiVersion(inputRequestData, buff))) // Error code
-	binary.BigEndian.PutUint16(buff[8:10], uint16(18))                                         // Api Version
+	binary.BigEndian.PutUint32(buff[8:10], uint32(validateApiVersion(inputRequestData, buff))) // Error code
+	binary.BigEndian.PutUint32(buff[8:10], uint32(18))                                         // Api Version
 	conn.Write(buff)
 }
